+++ benchmark_examples/autoattack/applications/image/cifar10/resnet20/cifar10_resnet20.py
@@ -21,7 +21,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 import torch.nn.init as init
 import torch.optim as optim
-from torchmetrics import Accuracy, Precision
+from torchmetrics import AUROC, Accuracy, Precision
 
 import secretflow as sf
 from benchmark_examples.autoattack import global_config
@@ -256,6 +256,7 @@ class Cifar10Resnet20(Cifar10ApplicationBase):
                 metric_wrapper(
                     Precision, task="multiclass", num_classes=10, average='micro'
                 ),
+                metric_wrapper(AUROC, task="multiclass", num_classes=10),
             ],
         )