+++ benchmark_examples/autoattack/defenses/de_identification.py
@@ -127,7 +127,7 @@ class DeIdentification(DefenseBase):
             and app.dataset_type() != DatasetType.IMAGE
         )
 
-    def tune_metrics(self) -> Dict[str, str]:
+    def tune_metrics(self, app_metrics: Dict[str, str]) -> Dict[str, str]:
         return {}
 
     def update_resources_consumptions(