+++ benchmark_examples/autoattack/defenses/fed_pass.py
@@ -74,7 +74,7 @@ class FedPass(DefenseBase):
             ModelType.DEEPFM,
         ]
 
-    def tune_metrics(self) -> Dict[str, str]:
+    def tune_metrics(self, app_metrics: Dict[str, str]) -> Dict[str, str]:
         return {}
 
     def update_resources_consumptions(