+++ benchmark_examples/autoattack/defenses/mixup.py
@@ -44,7 +44,7 @@ class Mixup(DefenseBase):
     def check_attack_valid(self, attack: AttackBase) -> bool:
         return attack.attack_type() == AttackType.LABLE_INFERENSE
 
-    def tune_metrics(self) -> Dict[str, str]:
+    def tune_metrics(self, app_metrics: Dict[str, str]) -> Dict[str, str]:
         return {}
 
     def check_app_valid(self, app: ApplicationBase) -> bool: