+++ secretflow/component/entry.py
@@ -41,7 +41,10 @@ from secretflow.component.preprocessing.binning.vert_woe_binning import (
     vert_woe_binning_comp,
 )
 from secretflow.component.preprocessing.data_prep.psi import psi_comp
+<<<<<<< HEAD
 from secretflow.component.preprocessing.filter.sample import sample_comp
+=======
+>>>>>>> main
 from secretflow.component.preprocessing.data_prep.train_test_split import (
     train_test_split_comp,
 )
@@ -49,7 +52,11 @@ from secretflow.component.preprocessing.data_prep.union import union_comp
 from secretflow.component.preprocessing.filter.condition_filter import (
     condition_filter_comp,
 )
+from secretflow.component.preprocessing.filter.expr_condition_filter import (
+    expr_condition_filter_comp,
+)
 from secretflow.component.preprocessing.filter.feature_filter import feature_filter_comp
+from secretflow.component.preprocessing.filter.sample import sample_comp
 from secretflow.component.preprocessing.unified_single_party_ops.binary_op import (
     binary_op_comp,
 )
@@ -91,6 +98,7 @@ ALL_COMPONENTS = [
     vert_woe_binning_comp,
     vert_bin_substitution_comp,
     condition_filter_comp,
+    expr_condition_filter_comp,
     ss_vif_comp,
     ss_pearsonr_comp,
     ss_pvalue_comp,
@@ -152,10 +160,12 @@ COMP_LIST, COMP_MAP = generate_comp_list()
 
 def get_comp_def(domain: str, name: str, version: str) -> ComponentDef:
     key = gen_key(domain, name, version)
-    assert (
-        key in COMP_MAP
-    ), f"key {key} is not in component list {list(COMP_MAP.keys())}"
-    return COMP_MAP[key].definition()
+    if key in COMP_MAP:
+        return COMP_MAP[key].definition()
+    else:
+        raise AttributeError(
+            f"key {key} is not in component list {list(COMP_MAP.keys())}"
+        )
 
 
 def comp_eval(
@@ -166,17 +176,19 @@ def comp_eval(
 ) -> NodeEvalResult:
     import logging
 
-    logging.warning(f"\n--\n{build_message()}\n--\n")
-    logging.warning(f'\n--\n*param* \n\n{param}\n--\n')
-    logging.warning(f'\n--\n*storage_config* \n\n{storage_config}\n--\n')
-    logging.warning(f'\n--\n*cluster_config* \n\n{cluster_config}\n--\n')
+    logging.info(f"\n--\n{build_message()}\n--\n")
+    logging.info(f'\n--\n*param* \n\n{param}\n--\n')
+    logging.info(f'\n--\n*storage_config* \n\n{storage_config}\n--\n')
+    logging.info(f'\n--\n*cluster_config* \n\n{cluster_config}\n--\n')
     key = gen_key(param.domain, param.name, param.version)
     if key in COMP_MAP:
         comp = COMP_MAP[key]
         res = comp.eval(
             param, storage_config, cluster_config, tracer_report=tracer_report
         )
-        logging.warning(f'\n--\n*res* \n\n{res}\n--\n')
+        logging.info(f'\n--\n*res* \n\n{res}\n--\n')
         return res
     else:
-        raise RuntimeError("component is not found.")
+        raise AttributeError(
+            f"key {key} is not in component list {list(COMP_MAP.keys())}"
+        )